Move all internal handlers to `Io` namespace
