+++ src/Io/RedirectHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace FrameworkX;
+namespace FrameworkX\Io;
 
 use Psr\Http\Message\ResponseInterface;
 use React\Http\Message\Response;
