+++ tests/AppTest.php
@@ -6,10 +6,10 @@ use FrameworkX\AccessLogHandler;
 use FrameworkX\App;
 use FrameworkX\Container;
 use FrameworkX\ErrorHandler;
-use FrameworkX\FiberHandler;
-use FrameworkX\MiddlewareHandler;
-use FrameworkX\RouteHandler;
-use FrameworkX\SapiHandler;
+use FrameworkX\Io\FiberHandler;
+use FrameworkX\Io\MiddlewareHandler;
+use FrameworkX\Io\RouteHandler;
+use FrameworkX\Io\SapiHandler;
 use FrameworkX\Tests\Fixtures\InvalidAbstract;
 use FrameworkX\Tests\Fixtures\InvalidConstructorInt;
 use FrameworkX\Tests\Fixtures\InvalidConstructorIntersection;
