+++ tests/Io/HtmlHandlerTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Framework\Tests;
+namespace Framework\Tests\Io;
 
+use FrameworkX\Io\HtmlHandler;
 use PHPUnit\Framework\TestCase;
-use FrameworkX\HtmlHandler;
 
 class HtmlHandlerTest extends TestCase
 {
